Update order row content

+++ ShoB/Home.swift
@@ -19,10 +19,10 @@ struct Home: View {
         dataSource.performFetch()
         
         // reset code
-        dataSource.fetchController.fetchedObjects?.forEach {
-            dataSource.context.delete($0)
-        }
-        dataSource.context.quickSave()
+//        dataSource.fetchController.fetchedObjects?.forEach {
+//            dataSource.context.delete($0)
+//        }
+//        dataSource.context.quickSave()
         
         return dataSource
     }()
@@ -54,27 +54,29 @@ struct Home: View {
         dataSource.performFetch()
         
         // reset code
-        dataSource.fetchController.fetchedObjects?.forEach {
-            dataSource.context.delete($0)
+        if dataSource.fetchController.fetchedObjects!.isEmpty {
+            dataSource.fetchController.fetchedObjects?.forEach {
+                dataSource.context.delete($0)
+            }
+            
+            let customer1 = Customer(context: dataSource.context)
+            customer1.givenName = "Tristana"
+            customer1.familyName = "Yordle"
+            customer1.organization = "Sumonner Rift"
+            customer1.contact.phone = "562 111 2222"
+            customer1.contact.email = "trist@email.com"
+            customer1.contact.address = "1234 Rift Ave, Long Beach, CA 90805"
+            
+            let customer2 = Customer(context: dataSource.context)
+            customer2.givenName = "Sivir"
+            customer2.familyName = "Surima"
+            customer2.organization = "Sand Empire"
+            customer2.contact.phone = "562 444 5555"
+            customer2.contact.email = "sivir@email.com"
+            customer2.contact.address = "1010 Sand Ave, Long Beach, CA 90805"
+            
+            dataSource.context.quickSave()
         }
-        
-        let customer1 = Customer(context: dataSource.context)
-        customer1.givenName = "Tristana"
-        customer1.familyName = "Yordle"
-        customer1.organization = "Sumonner Rift"
-        customer1.contact.phone = "562 111 2222"
-        customer1.contact.email = "trist@email.com"
-        customer1.contact.address = "1234 Rift Ave, Long Beach, CA 90805"
-        
-        let customer2 = Customer(context: dataSource.context)
-        customer2.givenName = "Sivir"
-        customer2.familyName = "Surima"
-        customer2.organization = "Sand Empire"
-        customer2.contact.phone = "562 444 5555"
-        customer2.contact.email = "sivir@email.com"
-        customer2.contact.address = "1010 Sand Ave, Long Beach, CA 90805"
-        
-        dataSource.context.quickSave()
         
         return dataSource
     }()
